Fix prepare:gapminder step
The prepare step is only necessary if a web product has additional dependencies or is a server product.
In our case the web product `gapminder` does not hat additional dependencies, so the preparation step can be omitted.
This fix checks if a workspace is needed and skips adding the prepare step to the chain otherwise.